[Refactor] The sequence parallelism characteristics in the MoE and Dense models are integrated into a single solution.
